Move helper functions from TraversalWrapper to Gremlin for reuse throughout
