+++ lib/gremlin.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var _ = require('lodash');
+var assert = require('assert');
 var fs = require('fs');
 var glob = require('glob');
 var path = require('path');
@@ -258,4 +259,141 @@ Gremlin.prototype.importGroovy = function (javaClassOrPkg) {
 // choose, etc.)
 Gremlin.prototype.newJavaScriptLambda = function (javascript) {
   return new this.ScriptEngineLambda(this._javaScriptEngineName, javascript);
+};
+
+Gremlin.prototype._parseVarargs = function (args, type) {
+  var va, self = this;
+  if (_.isArray(args[args.length - 1])) {
+    va = args.pop();
+  } else {
+    va = [];
+    // HACK - instead of actually converting JS strings -> java.lang.String
+    // instances as part of javify, we check the type with _isString
+    var test = type === 'java.lang.String' ? _.isString : function (o) {
+      return self.gremlin.isType(o, type);
+    };
+    while (test(args[args.length - 1])) {
+      va.unshift(args.pop());
+    }
+  }
+  args.push(this.java.newArray(type, va));
+};
+
+Gremlin.prototype._isClosure = function (val) {
+  var closureRegex = /^\{.*\}$/;
+  return _.isString(val) && val.search(closureRegex) > -1;
+};
+
+Gremlin.prototype._toGroovyLambda = function (groovyClosureString) {
+  var self = this;
+  assert.ok(self._isClosure(groovyClosureString));
+  var lambda = self.newGroovyLambda(groovyClosureString);
+  return lambda;
+};
+
+Gremlin.prototype._toGroovyLambdas = function (groovyClosureStrings) {
+  var self = this;
+  var lambdas = groovyClosureStrings.map(function (groovy) {
+    return self._toGroovyLambda(groovy);
+  });
+
+  // Put them in a Java array of GroovyLambda type.
+  var className = this.GroovyLambda.class.getNameSync();
+  var javaArray = this.java.newArray(className, lambdas);
+  return javaArray;
+};
+
+Gremlin.prototype._javify = function (arg) {
+  if (arg.unwrap) {
+    return arg.unwrap();
+  } else if (this._isClosure(arg)) {
+    // Turn something that looks like a Groovy closure into a GroovyLambda, which implements a variety of function
+    // interfaces.
+    var lambda = this._toGroovyLambda(arg);
+    return lambda;
+  }
+  return arg;
+};
+
+Gremlin.prototype._jsify = function (arg) {
+  if (!_.isObject(arg)) {
+    return arg;
+  }
+
+  if (!arg._isType) {
+    arg._isType = {};
+  }
+
+  if (arg.longValue) {
+    arg._isType.longValue = true;
+    return parseInt(arg.longValue, 10);
+  } else if (this.isType(arg, 'com.tinkerpop.gremlin.structure.Vertex')) {
+    return this.wrapVertex(arg);
+  } else if (this.isType(arg, 'com.tinkerpop.gremlin.structure.Edge')) {
+    return this.wrapEdge(arg);
+  } else if (this.isType(arg, 'java.util.List')) {
+    var arr = [];
+    var it = arg.iteratorSync();
+    while (it.hasNextSync()) {
+      var elem = it.nextSync();
+      var obj = this._jsify(elem);
+      arr.push(obj);
+    }
+    return arr;
+  } else if (this.isType(arg, 'java.util.Map')) {
+    // it seems this type of coercion could be ported to node-java
+    // https://github.com/joeferner/node-java/issues/56
+    var map = {};
+    var it = arg.entrySetSync().iteratorSync();
+    while (it.hasNextSync()) {
+      var pair = it.nextSync();
+      map[pair.getKeySync()] = this._jsify(pair.getValueSync());
+    }
+    return map;
+  } else if (this.isType(arg, 'com.tinkerpop.gremlin.process.util.BulkSet')) {
+    var arr = [];
+    var it = arg.iteratorSync();
+    while (it.hasNextSync()) {
+      var key = it.nextSync();
+      var count = this._jsify(arg.getSync(key));
+      var obj = {
+        key: this._jsify(key),
+        count: count
+      };
+      arr.push(obj);
+    }
+    return arr;
+  } else if (this.isType(arg, 'java.lang.Object')) {
+    return arg;
+  }
+  return arg;
+};
+
+Gremlin.prototype._asJSON = function (elem) {
+  var self = this;
+  var java = this.java;
+  if (!_.isObject(elem)) {
+    // Scalars should stay that way.
+    return elem;
+
+  } else if (_.isArray(elem)) {
+    // Arrays must be recursively converted.
+    return elem.map(function (e) { return self._asJSON(e); });
+
+  } else if (java.instanceOf(elem, 'java.lang.Object')) {
+    // If we still have an unrecognized Java object, convert it to a string.
+    return {'javaClass': elem.getClassSync().getNameSync(), 'toString': elem.toStringSync()};
+
+  } else if ('toJSON' in elem) {
+    // If we have a 'toJSON' method, use it.
+    return elem;
+
+  } else {
+    // Recursively convert any other kind of object.
+    var json = {};
+    _.forEach(elem, function (value, key) {
+      json[key] = self._asJSON(value);
+    });
+    return json;
+  }
 };
