+++ lib/traversal-wrapper.js
@@ -16,114 +16,6 @@ var TraversalWrapper = module.exports = function (gremlin, traversal) {
   this.traversal = traversal;
 };
 
-TraversalWrapper.prototype._parseVarargs = function (args, type) {
-  var va, self = this;
-  if (_.isArray(args[args.length - 1])) {
-    va = args.pop();
-  } else {
-    va = [];
-    // HACK - instead of actually converting JS strings -> java.lang.String
-    // instances as part of javify, we check the type with _isString
-    var test = type === 'java.lang.String' ? _.isString : function (o) {
-      return self.gremlin.isType(o, type);
-    };
-    while (test(args[args.length - 1])) {
-      va.unshift(args.pop());
-    }
-  }
-  args.push(this.gremlin.java.newArray(type, va));
-};
-
-TraversalWrapper.prototype._isClosure = function (val) {
-  var closureRegex = /^\{.*\}$/;
-  return _.isString(val) && val.search(closureRegex) > -1;
-};
-
-TraversalWrapper.prototype._toGroovyLambda = function (groovyClosureString) {
-  var self = this;
-  assert.ok(self._isClosure(groovyClosureString));
-  var lambda = self.gremlin.newGroovyLambda(groovyClosureString);
-  return lambda;
-};
-
-TraversalWrapper.prototype._toGroovyLambdas = function (groovyClosureStrings) {
-  var self = this;
-  var lambdas = groovyClosureStrings.map(function (groovy) {
-    return self._toGroovyLambda(groovy);
-  });
-
-  // Put them in a Java array of GroovyLambda type.
-  var className = this.gremlin.GroovyLambda.class.getNameSync();
-  var javaArray = this.gremlin.java.newArray(className, lambdas);
-  return javaArray;
-};
-
-TraversalWrapper.prototype._javify = function (arg) {
-  if (arg.unwrap) {
-    return arg.unwrap();
-  } else if (this._isClosure(arg)) {
-    // Turn something that looks like a Groovy closure into a GroovyLambda, which implements a variety of function
-    // interfaces.
-    var lambda = this._toGroovyLambda(arg);
-    return lambda;
-  }
-  return arg;
-};
-
-TraversalWrapper.prototype._jsify = function (arg) {
-  if (!_.isObject(arg)) {
-    return arg;
-  }
-
-  if (!arg._isType) {
-    arg._isType = {};
-  }
-
-  if (arg.longValue) {
-    arg._isType.longValue = true;
-    return parseInt(arg.longValue, 10);
-  } else if (this.gremlin.isType(arg, 'com.tinkerpop.gremlin.structure.Vertex')) {
-    return this.gremlin.wrapVertex(arg);
-  } else if (this.gremlin.isType(arg, 'com.tinkerpop.gremlin.structure.Edge')) {
-    return this.gremlin.wrapEdge(arg);
-  } else if (this.gremlin.isType(arg, 'java.util.List')) {
-    var arr = [];
-    var it = arg.iteratorSync();
-    while (it.hasNextSync()) {
-      var elem = it.nextSync();
-      var obj = this._jsify(elem);
-      arr.push(obj);
-    }
-    return arr;
-  } else if (this.gremlin.isType(arg, 'java.util.Map')) {
-    // it seems this type of coercion could be ported to node-java
-    // https://github.com/joeferner/node-java/issues/56
-    var map = {};
-    var it = arg.entrySetSync().iteratorSync();
-    while (it.hasNextSync()) {
-      var pair = it.nextSync();
-      map[pair.getKeySync()] = this._jsify(pair.getValueSync());
-    }
-    return map;
-  } else if (this.gremlin.isType(arg, 'com.tinkerpop.gremlin.process.util.BulkSet')) {
-    var arr = [];
-    var it = arg.iteratorSync();
-    while (it.hasNextSync()) {
-      var key = it.nextSync();
-      var count = this._jsify(arg.getSync(key));
-      var obj = {
-        key: this._jsify(key),
-        count: count
-      };
-      arr.push(obj);
-    }
-    return arr;
-  } else if (this.gremlin.isType(arg, 'java.lang.Object')) {
-    return arg;
-  }
-  return arg;
-};
-
 TraversalWrapper.prototype.unwrap = function () {
   return this.traversal;
 };
@@ -167,13 +59,13 @@ TraversalWrapper.prototype.between = function () {
 
 TraversalWrapper.prototype.bothE = function () {
   var args = Array.prototype.slice.call(arguments);
-  this._parseVarargs(args, 'java.lang.String');
+  this.gremlin._parseVarargs(args, 'java.lang.String');
   return this.add('bothE', args);
 };
 
 TraversalWrapper.prototype.both = function () {
   var args = Array.prototype.slice.call(arguments);
-  this._parseVarargs(args, 'java.lang.String');
+  this.gremlin._parseVarargs(args, 'java.lang.String');
   return this.add('both', args);
 };
 
@@ -197,13 +89,13 @@ TraversalWrapper.prototype.idVertex = function () {
 
 TraversalWrapper.prototype.inE = function () {
   var args = Array.prototype.slice.call(arguments);
-  this._parseVarargs(args, 'java.lang.String');
+  this.gremlin._parseVarargs(args, 'java.lang.String');
   return this.add('inE', args);
 };
 
 TraversalWrapper.prototype.in = function () {
   var args = Array.prototype.slice.call(arguments);
-  this._parseVarargs(args, 'java.lang.String');
+  this.gremlin._parseVarargs(args, 'java.lang.String');
   return this.add('in', args);
 };
 
@@ -237,13 +129,13 @@ TraversalWrapper.prototype.label = function () {
 
 TraversalWrapper.prototype.outE = function () {
   var args = Array.prototype.slice.call(arguments);
-  this._parseVarargs(args, 'java.lang.String');
+  this.gremlin._parseVarargs(args, 'java.lang.String');
   return this.add('outE', args);
 };
 
 TraversalWrapper.prototype.out = function () {
   var args = Array.prototype.slice.call(arguments);
-  this._parseVarargs(args, 'java.lang.String');
+  this.gremlin._parseVarargs(args, 'java.lang.String');
   return this.add('out', args);
 };
 
@@ -253,7 +145,7 @@ TraversalWrapper.prototype.outV = function () {
 
 TraversalWrapper.prototype.map = function () {
   var args = Array.prototype.slice.call(arguments);
-  this._parseVarargs(args, 'java.lang.String');
+  this.gremlin._parseVarargs(args, 'java.lang.String');
   return this.add('map', args);
 };
 
@@ -264,12 +156,12 @@ TraversalWrapper.prototype.value = function () {
 
 TraversalWrapper.prototype.values = function () {
   var args = Array.prototype.slice.call(arguments);
-  this._parseVarargs(args, 'java.lang.String');
+  this.gremlin._parseVarargs(args, 'java.lang.String');
   return this.add('values', args);
 };
 
 TraversalWrapper.prototype.step = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   return this.add('step', args);
 };
 
@@ -278,7 +170,7 @@ TraversalWrapper.prototype.step = function () {
 /////////////////////////
 
 TraversalWrapper.prototype.choose = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   // Two overloads distinguished by number of args:
   // choose(Predicate<E> choose, Traversal<?, E2> trueChoice, Traversal<?, E2> falseChoice)
   // choose(Function<E, M> map, Map<M, Traversal> choices)
@@ -307,8 +199,8 @@ TraversalWrapper.prototype.choose = function () {
 };
 
 TraversalWrapper.prototype.copySplit = function () {  // TK2 only?
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
-  this._parseVarargs(args, 'com.tinkerpop.gremlin.process.Traversal');
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
+  this.gremlin._parseVarargs(args, 'com.tinkerpop.gremlin.process.Traversal');
   return this.add('copySplit', args);
 };
 
@@ -321,7 +213,7 @@ TraversalWrapper.prototype.fairMerge = function () {
 };
 
 TraversalWrapper.prototype.ifThenElse = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   return this.add('ifThenElse', args);
 };
 
@@ -330,8 +222,8 @@ TraversalWrapper.prototype.ifThenElse = function () {
 /////////////////////////
 
 TraversalWrapper.prototype.and = function (/*final Traversal<E, ?>... traversals*/) {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
-  this._parseVarargs(args, 'com.tinkerpop.gremlin.process.Traversal');
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
+  this.gremlin._parseVarargs(args, 'com.tinkerpop.gremlin.process.Traversal');
   return this.add('and', args);
 };
 
@@ -341,7 +233,7 @@ TraversalWrapper.prototype.back = function (step) {
 };
 
 TraversalWrapper.prototype.dedup = function (closure) {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   return this.add('dedup', args);
 };
 
@@ -354,19 +246,19 @@ TraversalWrapper.prototype.except = function () {
     args[0] = this.gremlin.toListSync(args[0]);
   } else {
     // assume except(final String... namedSteps)
-    this._parseVarargs(args, 'java.lang.String');
+    this.gremlin._parseVarargs(args, 'java.lang.String');
   }
   return this.add('except', args);
 };
 
 TraversalWrapper.prototype.filter = function (closure) {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   return this.add('filter', args);
 };
 
 TraversalWrapper.prototype.or = function (/*final Traversal<E, ?>... traversals*/) {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
-  this._parseVarargs(args, 'com.tinkerpop.gremlin.process.Traversal');
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
+  this.gremlin._parseVarargs(args, 'com.tinkerpop.gremlin.process.Traversal');
   return this.add('or', args);
 };
 
@@ -393,7 +285,7 @@ TraversalWrapper.prototype.retain = function (/*final Collection<E> collection*/
     args[0] = this.gremlin.toListSync(args[0]);
   } else {
     // assume retain(final String... namedSteps)
-    this._parseVarargs(args, 'java.lang.String');
+    this.gremlin._parseVarargs(args, 'java.lang.String');
   }
   return this.add('retain', args);
 };
@@ -407,7 +299,7 @@ TraversalWrapper.prototype.simplePath = function () {
 //////////////////////////////
 
 // TraversalWrapper.prototype.aggregate = function () {
-//   var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+//   var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
 //   if (_.isArray(args[0])) {
 //     args[0] = this.gremlin.toListSync(args[0]);
 //   }
@@ -425,17 +317,17 @@ TraversalWrapper.prototype.optional = function () {
 };
 
 TraversalWrapper.prototype.groupBy = function (map, closure) {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   return this.add('groupBy', args);
 };
 
 TraversalWrapper.prototype.groupCount = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   return this.add('groupCount', args);
 };
 
 TraversalWrapper.prototype.inject = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   return this.add('inject', args);
 };
 
@@ -455,17 +347,17 @@ TraversalWrapper.prototype.linkBoth = function () {
 };
 
 TraversalWrapper.prototype.sideEffect = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   return this.add('sideEffect', args);
 };
 
 TraversalWrapper.prototype.sack = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   return this.add('sack', args);
 };
 
 TraversalWrapper.prototype.store = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   if (_.isArray(args[0])) {
     args[0] = this.gremlin.toListSync(args[0]);
   }
@@ -473,8 +365,8 @@ TraversalWrapper.prototype.store = function () {
 };
 
 TraversalWrapper.prototype.table = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
-  this._parseVarargs(args, 'groovy.lang.Closure');
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
+  this.gremlin._parseVarargs(args, 'groovy.lang.Closure');
   if (_.isArray(args[1])) {
     args[1] = this.gremlin.toListSync(args[1]);
   }
@@ -482,13 +374,13 @@ TraversalWrapper.prototype.table = function () {
 };
 
 TraversalWrapper.prototype.tree = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
-  this._parseVarargs(args, 'groovy.lang.Closure');
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
+  this.gremlin._parseVarargs(args, 'groovy.lang.Closure');
   return this.add('tree', args);
 };
 
 TraversalWrapper.prototype.withSack = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   return this.add('withSack', args);
 };
 
@@ -497,7 +389,7 @@ TraversalWrapper.prototype.withSack = function () {
 ////////////////////////////
 
 TraversalWrapper.prototype.gather = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   return this.add('gather', args);
 };
 
@@ -511,13 +403,12 @@ TraversalWrapper.prototype.memoize = function () {
 };
 
 TraversalWrapper.prototype.order = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   return this.add('order', args);
 };
 
 TraversalWrapper.prototype.path = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
-  this._parseVarargs(args, 'groovy.lang.Closure');
+  var args = Array.prototype.slice.call(arguments);
   return this.add('path', args);
 };
 
@@ -533,7 +424,7 @@ TraversalWrapper.prototype.scatter = function () {
 TraversalWrapper.prototype.select = function () {
   var args = Array.prototype.slice.call(arguments);
   if (args.length !== 1) {
-    this._parseVarargs(args, 'java.lang.String');
+    this.gremlin._parseVarargs(args, 'java.lang.String');
   }
   return this.add('select', args);
 };
@@ -552,12 +443,12 @@ TraversalWrapper.prototype.count = function () {
 };
 
 TraversalWrapper.prototype.orderMap = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   return this.add('orderMap', args);
 };
 
 TraversalWrapper.prototype.transform = function () {
-  var args = Array.prototype.slice.call(arguments).map(this._javify.bind(this));
+  var args = Array.prototype.slice.call(arguments).map(this.gremlin._javify.bind(this.gremlin));
   return this.add('transform', args);
 };
 
@@ -596,7 +487,7 @@ function traversalPromiseJsifyWrap(op) {
     dlog('traversalPromiseJsifyWrap(%s)', op, argPair, this.traversal[op]);
     return Q.npost(this.traversal, op, argPair.args)
       .then(function (res) {
-        var jres = self._jsify(res);
+        var jres = self.gremlin._jsify(res);
         dlog('traversalPromiseJsifyWrap(%s)(result) = ', res, jres);
         return jres;
       })
@@ -630,7 +521,7 @@ TraversalWrapper.prototype.toArray = function (callback) {
       var arr = [];
       for (var i = 0, l = list.sizeSync(); i < l; i++) {
         var it = list.getSync(i);
-        arr.push(self._jsify(it));
+        arr.push(self.gremlin._jsify(it));
       }
       dlog('TraversalWrapper.prototype.toArray:', arr.length, arr);
       return arr;
@@ -644,45 +535,16 @@ TraversalWrapper.prototype.toArraySync = function () {
   var arr = [];
   for (var i = 0, l = list.sizeSync(); i < l; i++) {
     var it = list.getSync(i);
-    arr.push(self._jsify(it));
+    arr.push(self.gremlin._jsify(it));
   }
   dlog('TraversalWrapper.prototype.toArraySync:', arr.length, arr);
   return arr;
 };
 
-TraversalWrapper.prototype._asJSON = function (elem) {
-  var self = this;
-  var java = this.gremlin.java;
-  if (!_.isObject(elem)) {
-    // Scalars should stay that way.
-    return elem;
-
-  } else if (_.isArray(elem)) {
-    // Arrays must be recursively converted.
-    return elem.map(function (e) { return self._asJSON(e); });
-
-  } else if (java.instanceOf(elem, 'java.lang.Object')) {
-    // If we still have an unrecognized Java object, convert it to a string.
-    return {'javaClass': elem.getClassSync().getNameSync(), 'toString': elem.toStringSync()};
-
-  } else if ('toJSON' in elem) {
-    // If we have a 'toJSON' method, use it.
-    return elem;
-
-  } else {
-    // Recursively convert any other kind of object.
-    var json = {};
-    _.forEach(elem, function (value, key) {
-      json[key] = self._asJSON(value);
-    });
-    return json;
-  }
-};
-
 TraversalWrapper.prototype.asJSONSync = function () {
   var self = this;
   var array = this.toArraySync().map(function (elem) {
-    return self._asJSON(elem);
+    return self.gremlin._asJSON(elem);
   });
 
   return JSON.parse(JSON.stringify(array));
@@ -703,7 +565,7 @@ TraversalWrapper.prototype.forEach = function (process, callback) {
 // Here we automatically call `next()` so that we can return a `GraphWrapper` instance for the subgraph.
 TraversalWrapper.prototype.subgraph = function (edgePredicate, callback) {
   var self = this;
-  edgePredicate = self._javify(edgePredicate);
+  edgePredicate = self.gremlin._javify(edgePredicate);
   return Q.nbind(self.traversal.subgraph, self.traversal)(edgePredicate)
     .then(function (trav) { return self.gremlin.wrapTraversal(trav).next(); })
     .then(function (sg) { return new GraphWrapper(self.gremlin, sg); })
